+++ src/ui.rs
@@ -11,6 +11,7 @@ mod dialog_archive_prefix;
 mod dialog_device_missing;
 mod dialog_encryption_password;
 mod dialog_exclude;
+mod dialog_exclude_pattern;
 mod dialog_info;
 mod dialog_prune;
 mod dialog_prune_review;